requirements.txt
Flask==2.1.2
werkzeug==2.0.3
pymongo[srv]
gunicorn==20.1.0
pymongo>=4.1.1 
flask_cors
python-dotenv